refactor: rename utils package, make public
