+++ packages/cli-plugin-utils/README.md
@@ -1,3 +1,3 @@
-# `@vuetify/shared`
+# `@vuetify/cli-plugin-utils`
 
 > This package provides shared utility helper functions for all plugin packages
