+++ packages/vue-cli-plugin-vuetify-preset-crane/generator.js
@@ -1,7 +1,7 @@
 const {
   addHtmlLink,
   generatePreset,
-} = require('@vuetify/shared')
+} = require('@vuetify/cli-plugin-utils')
 
 module.exports = api => {
   generatePreset(api, 'crane', () => {
