+++ packages/vue-cli-plugin-vuetify-preset-crane/index.js
@@ -1,4 +1,4 @@
 // Imports
-const { bootstrapPreset } = require('@vuetify/shared')
+const { bootstrapPreset } = require('@vuetify/cli-plugin-utils')
 
 module.exports = api => bootstrapPreset(api, 'crane')
